+++ flexiznam/cli.py
@@ -66,6 +66,42 @@ def add_password(app, username, password, password_file):
 
 
 @cli.command()
+@click.option('-s', '--source_dir', required=True,
+              help='Base directory for this yaml file. Usually a session directory')
+@click.option('-t', '--target_yaml', required=True, help='Path to output YAML file.')
+@click.option('-p', '--project', default=None, help='Project name on flexilims.')
+@click.option('-m', '--mouse', default=None, help='Mouse name on flexilims.')
+@click.option('--overwrite/--no-overwrite', default=False,
+              help='If target yaml exist, should I replace it?.')
+@click.option('--process/--no-process', default=False,
+              help='After creating the yaml skeleton, should I also parse it?')
+@click.option('-r', '--raw_data_folder', default=None,
+              help='Path to the root folder containing raw data. Only used with '
+                   '`--process`')
+def create_yaml(source_dir, target_yaml, project, mouse, overwrite, process,
+                raw_data_folder):
+    """Create a yaml file by looking recursively in `root_dir`"""
+
+    target_yaml = pathlib.Path(target_yaml)
+    if (not overwrite) and target_yaml.exists():
+        s = input('File %s already exists. Overwrite (yes/[no])? ' % target_yaml)
+        if s == 'yes':
+            overwrite = True
+        else:
+            raise(FileExistsError('File %s already exists and overwrite is not allowed'
+                                  % target_yaml))
+    source_dir = pathlib.Path(source_dir)
+    if not source_dir.is_dir():
+        raise FileNotFoundError('source_dir %s is not a directory' % source_dir)
+    yml_content = camp.sync_data.create_yaml(root_folder=source_dir,
+                                             outfile=target_yaml,
+                                             project=project,
+                                             mouse=mouse,
+                                             overwrite=overwrite)
+    click.echo('Created yml skeleton in %s' % target_yaml)
+
+
+@cli.command()
 @click.option('-s', '--source_yaml', required=True, help='Manually generated yaml to seed automatic method.')
 @click.option('-t', '--target_yaml', default=None, help='Path to outpout YAML file.')
 @click.option('-r', '--raw_data_folder', default=None, help='Path to the root folder containing raw data')